LW1/LW1.4 updated

+++ LW1/LW1.4/include/calc_gamma.h
@@ -1,11 +1,7 @@
 #pragma once
 
-#include <stdio.h>
-#include <math.h>
-#include <stdlib.h>
-
 #include "../include/calc_pi.h"
 
-double calc_gamma_lim(const double eps);
-double calc_gamma_row(const double eps, const double pi);
-double calc_gamma_equation(const double eps);
+long double calc_gamma_lim(const long double eps);
+long double calc_gamma_row(const long double eps, const long double pi);
+long double calc_gamma_equation(const long double eps);
