+++ LW1/LW1.4/src/calculate_eps.c
@@ -1,6 +1,6 @@
 #include "../include/calculate_eps.h"
 
-double calculate_epsilon(const int precision)
+long double calculate_epsilon(const int precision)
 {   
     double eps = 1.0;
     for (int i = precision; i > 0; --i)
